qualifier/requirements.txt
pyssim
Pillow
zerorpc